Enforce fallback support for float vector retrieval in quantized KNN vector formats (#15476)
* Moved testReadQuantizedVectorWithEmptyRawVectors test case to BaserKnnVectorsFormatTestCase

* Added entry to Changes.txt

* Rename function

* Made helper functions protected

* CHanged function name and remove redundanct vector copying

* Moved helper methods to subclasses to have codec specific implementation

* Revert "Moved helper methods to subclasses to have codec specific implementation"

This reverts commit f0c66d6e3060def04a7a6ef68c0eacf41ee6c0e8.

* Moved helper methods to subclasses to have codec specific implementation which supports fallback

* remove unused constant

* Made functions private from protected

* Moved indexWriter to try block

---------
