+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene102/TestLucene102BinaryQuantizedVectorsFormat.java
@@ -186,4 +186,9 @@ public class TestLucene102BinaryQuantizedVectorsFormat extends BaseKnnVectorsFor
       }
     }
   }
+
+  @Override
+  protected boolean supportsFloatVectorFallback() {
+    return false;
+  }
 }
