+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene102/TestLucene102HnswBinaryQuantizedVectorsFormat.java
@@ -180,4 +180,9 @@ public class TestLucene102HnswBinaryQuantizedVectorsFormat extends BaseKnnVector
       }
     }
   }
+
+  @Override
+  protected boolean supportsFloatVectorFallback() {
+    return false;
+  }
 }
